feat: add seed parameter for reproducibility in charge sensor compensation
- Introduced a `seed` parameter in the `run_compensation` function to allow for reproducible measurement ordering by seeding the random number generator.
- Updated documentation to reflect the new parameter and its purpose, enhancing clarity for users.
- Removed outdated test function for inlier/outlier marking and added a new test to validate the RANSAC fitting process, ensuring comprehensive coverage of measurement sample handling.
